Separate phylogenetic methods into trees.py
